+++ ls/options/format.h
@@ -1,11 +1,9 @@
 #ifndef FORMAT_H_
 #define FORMAT_H_
 
-#include "../alias.h"
-
 void format(char **file_list, const char *full_path,
                               char **fmt_files,
-                              byte cnt);
+                              int cnt);
 
 #endif
